refactor(examples): renames the md files to README.md
this way, github displays them when navigating to the folder in their web ui

refs RBCOM-841
