+++ src/examples/panel-reusing/README.md
